+++ packages/common/web/src/Utils/normalization/dataNormalizers/index.js
@@ -1,2 +1,3 @@
 export { default as imageDataNormalizer } from './imageDataNormalizer';
 export { default as linkDataNormalizer } from './linkDataNormalizer';
+export { default as galleryDataNormalizer } from './galleryDataNormalizer';